# src/main/java/com/hdfclife/smartauth/dto/request/LoginRequest.java
package com.hdfclife.smartauth.dto.request;

import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank(message = "Username is required") String username,
        @NotBlank(message = "Password is required") String password) {}